[Datasets] Group file metadata providers in a single file and clarify docstrings.
