+++ python/ray/data/datasource/file_based_datasource.py
@@ -26,7 +26,10 @@ from ray.data.impl.arrow_block import ArrowRow
 from ray.data.impl.block_list import BlockMetadata
 from ray.data.impl.output_buffer import BlockOutputBuffer
 from ray.data.datasource.datasource import Datasource, ReadTask, WriteResult
-from ray.data.datasource.file_meta_provider import FileMetadataProvider
+from ray.data.datasource.file_meta_provider import (
+    BaseFileMetadataProvider,
+    DefaultFileMetadataProvider,
+)
 from ray.util.annotations import DeveloperAPI
 from ray.data.impl.util import _check_pyarrow_version
 from ray.data.impl.remote_fn import cached_remote_fn
@@ -118,156 +121,6 @@ class DefaultBlockWritePathProvider(BlockWritePathProvider):
         # FileSystem paths are always forward slash separated, see:
         # https://arrow.apache.org/docs/python/filesystems.html
         return posixpath.join(base_path, suffix)
-
-
-@DeveloperAPI
-class BaseFileMetadataProvider(FileMetadataProvider):
-    """Abstract callable that provides metadata for FileBasedDatasource
-     implementations that reuse the base `prepare_read` method.
-
-    Also supports file and file size discovery in input directory paths.
-
-     Current subclasses:
-         DefaultFileMetadataProvider
-    """
-
-    def _get_block_metadata(
-        self,
-        paths: List[str],
-        schema: Optional[Union[type, "pyarrow.lib.Schema"]],
-        *,
-        rows_per_file: Optional[int],
-        file_sizes: List[Optional[int]],
-    ) -> BlockMetadata:
-        """Resolves and returns block metadata for the given file paths.
-
-        Args:
-            paths: The file paths to aggregate block metadata across. These
-                paths will always be a subset of those previously returned from
-                `expand_paths()`.
-            schema: The user-provided or inferred schema for the given file
-                paths, if any.
-            rows_per_file: The fixed number of rows per input file, or None.
-            file_sizes: Optional file size per input file previously returned
-                from `expand_paths()`, where `file_sizes[i]` holds the size of
-                the file at `paths[i]`.
-
-        Returns:
-            BlockMetadata aggregated across the given file paths.
-        """
-        raise NotImplementedError
-
-    def expand_paths(
-        self,
-        paths: List[str],
-        filesystem: Optional["pyarrow.fs.FileSystem"],
-    ) -> Tuple[List[str], List[Optional[int]]]:
-        """Expands all paths into concrete file paths by walking directories.
-
-         Also returns a sidecar of file sizes.
-
-        The input paths must be normalized for compatibility with the input
-        filesystem prior to invocation.
-
-         Args:
-             paths: A list of file and/or directory paths compatible with the
-                 given filesystem.
-             filesystem: The filesystem implementation that should be used for
-                 expanding all paths and reading their files.
-
-         Returns:
-             A tuple whose first item contains the list of file paths discovered,
-             and whose second item contains the size of each file. `None` may be
-             returned if a file size is either unknown or will be fetched later
-             by `_get_block_metadata()`, but the length of both lists must be
-             equal.
-        """
-        raise NotImplementedError
-
-
-class DefaultFileMetadataProvider(BaseFileMetadataProvider):
-    """Default metadata provider for FileBasedDatasource implementations that
-    reuse the base `prepare_read` method.
-
-    Calculates block size in bytes as the sum of its constituent file sizes,
-    and assumes a fixed number of rows per file.
-    """
-
-    def _get_block_metadata(
-        self,
-        paths: List[str],
-        schema: Optional[Union[type, "pyarrow.lib.Schema"]],
-        *,
-        rows_per_file: Optional[int],
-        file_sizes: List[Optional[int]],
-    ) -> BlockMetadata:
-        if rows_per_file is None:
-            num_rows = None
-        else:
-            num_rows = len(paths) * rows_per_file
-        return BlockMetadata(
-            num_rows=num_rows,
-            size_bytes=None if None in file_sizes else sum(file_sizes),
-            schema=schema,
-            input_files=paths,
-            exec_stats=None,
-        )  # Exec stats filled in later.
-
-    def expand_paths(
-        self,
-        paths: List[str],
-        filesystem: "pyarrow.fs.FileSystem",
-    ) -> Tuple[List[str], List[Optional[int]]]:
-        from pyarrow.fs import FileType
-
-        if len(paths) > 1:
-            logger.warning(
-                f"Expanding {len(paths)} path(s). This may be a HIGH LATENCY "
-                f"operation on some cloud storage services. If the specified paths "
-                f"all point to files and never directories, try rerunning this read "
-                f"with `meta_provider=FastFileMetadataProvider()`."
-            )
-        expanded_paths = []
-        file_infos = []
-        for path in paths:
-            file_info = filesystem.get_file_info(path)
-            if file_info.type == FileType.Directory:
-                paths, file_infos_ = _expand_directory(path, filesystem)
-                expanded_paths.extend(paths)
-                file_infos.extend(file_infos_)
-            elif file_info.type == FileType.File:
-                expanded_paths.append(path)
-                file_infos.append(file_info)
-            else:
-                raise FileNotFoundError(path)
-        file_sizes = [file_info.size for file_info in file_infos]
-        return expanded_paths, file_sizes
-
-
-class FastFileMetadataProvider(DefaultFileMetadataProvider):
-    """Fast Metadata provider for FileBasedDatasource implementations.
-
-    Offers improved performance vs. DefaultFileMetadataProvider by skipping directory
-    path expansion and file size collection. While this performance improvement may be
-    negligible for local filesystems, it can be substantial for cloud storage service
-    providers.
-
-    This should only be used when all input paths are known to be files.
-    """
-
-    def expand_paths(
-        self,
-        paths: List[str],
-        filesystem: "pyarrow.fs.FileSystem",
-    ) -> Tuple[List[str], List[Optional[int]]]:
-        logger.warning(
-            f"Skipping expansion of {len(paths)} path(s). If your paths contain "
-            f"directories or if file size collection is required, try rerunning this "
-            f"read with `meta_provider=DefaultFileMetadataProvider()`."
-        )
-        import numpy as np
-
-        return paths, np.empty(len(paths), dtype=object)
 
 
 @DeveloperAPI
