+++ python/ray/data/datasource/parquet_datasource.py
@@ -1,9 +1,8 @@
 import logging
 import itertools
-from typing import Any, Callable, Optional, List, Union, Iterator, TYPE_CHECKING
+from typing import Callable, Optional, List, Union, Iterator, TYPE_CHECKING
 
 import numpy as np
-from ray.util.annotations import DeveloperAPI
 
 if TYPE_CHECKING:
     import pyarrow
@@ -15,12 +14,15 @@ from ray.data.context import DatasetContext
 from ray.data.datasource.datasource import ReadTask
 from ray.data.datasource.file_based_datasource import _resolve_paths_and_filesystem
 from ray.data.datasource.parquet_base_datasource import ParquetBaseDatasource
-from ray.data.datasource.file_meta_provider import FileMetadataProvider
-from ray.data.impl.block_list import BlockMetadata
+from ray.data.datasource.file_meta_provider import (
+    ParquetMetadataProvider,
+    DefaultParquetMetadataProvider,
+)
 from ray.data.impl.output_buffer import BlockOutputBuffer
 from ray.data.impl.progress_bar import ProgressBar
 from ray.data.impl.remote_fn import cached_remote_fn
 from ray.data.impl.util import _check_pyarrow_version
+
 
 logger = logging.getLogger(__name__)
 
@@ -51,115 +53,6 @@ def _deregister_parquet_file_fragment_serialization():
     from pyarrow.dataset import ParquetFileFragment
 
     ray.util.deregister_serializer(ParquetFileFragment)
-
-
-@DeveloperAPI
-class ParquetMetadataProvider(FileMetadataProvider):
-    """Abstract callable that provides metadata for Arrow Parquet file fragments.
-
-    Supports optional pre-fetching of ordered metadata for all file fragments in
-    a single batch to help optimize metadata resolution.
-
-    Current subclasses:
-        DefaultParquetMetadataProvider
-    """
-
-    def _get_block_metadata(
-        self,
-        paths: List[str],
-        schema: Optional[Union[type, "pyarrow.lib.Schema"]],
-        *,
-        pieces: List["pyarrow.dataset.ParquetFileFragment"],
-        prefetched_metadata: Optional[List[Any]],
-    ) -> BlockMetadata:
-        """Resolves and returns block metadata for the given file paths.
-
-        Args:
-            paths: The file paths to aggregate block metadata across.
-            schema: The user-provided or inferred schema for the given file
-                paths, if any.
-            pieces: The Parquet file fragments derived from the input file paths.
-            prefetched_metadata: Metadata previously returned from
-                `prefetch_file_metadata()` for each file fragment, where
-                `prefetched_metadata[i]` contains the metadata for `pieces[i]`.
-
-        Returns:
-            BlockMetadata aggregated across the given file paths.
-        """
-        raise NotImplementedError
-
-    def prefetch_file_metadata(
-        self,
-        pieces: List["pyarrow.dataset.ParquetFileFragment"],
-    ) -> Optional[List[Any]]:
-        """Pre-fetches file metadata for all Parquet file fragments in a single batch.
-
-        Subsets of the metadata returned will be provided as input to
-        subsequent calls to _get_block_metadata() together with their
-        corresponding Parquet file fragments.
-
-        Implementations that don't support pre-fetching file metadata shouldn't
-        override this method.
-
-        Args:
-            pieces: The Parquet file fragments to fetch metadata for.
-
-        Returns:
-            Metadata resolved for each input file fragment, or `None`. Metadata
-            must be returned in the same order as all input file fragments, such
-            that `metadata[i]` always contains the metadata for `pieces[i]`.
-        """
-        return None
-
-
-class DefaultParquetMetadataProvider(ParquetMetadataProvider):
-    """The default file metadata provider for ParquetDatasource.
-
-    Aggregates total block bytes and number of rows using the Parquet file metadata
-    associated with a list of Arrow Parquet dataset file fragments.
-    """
-
-    def _get_block_metadata(
-        self,
-        paths: List[str],
-        schema: Optional[Union[type, "pyarrow.lib.Schema"]],
-        *,
-        pieces: List["pyarrow.dataset.ParquetFileFragment"],
-        prefetched_metadata: Optional[List["pyarrow.parquet.FileMetaData"]],
-    ) -> BlockMetadata:
-        if prefetched_metadata is not None and len(prefetched_metadata) == len(pieces):
-            # Piece metadata was available, construct a normal
-            # BlockMetadata.
-            block_metadata = BlockMetadata(
-                num_rows=sum(m.num_rows for m in prefetched_metadata),
-                size_bytes=sum(
-                    sum(m.row_group(i).total_byte_size for i in range(m.num_row_groups))
-                    for m in prefetched_metadata
-                ),
-                schema=schema,
-                input_files=paths,
-                exec_stats=None,
-            )  # Exec stats filled in later.
-        else:
-            # Piece metadata was not available, construct an empty
-            # BlockMetadata.
-            block_metadata = BlockMetadata(
-                num_rows=None,
-                size_bytes=None,
-                schema=schema,
-                input_files=paths,
-                exec_stats=None,
-            )
-        return block_metadata
-
-    def prefetch_file_metadata(
-        self,
-        pieces: List["pyarrow.dataset.ParquetFileFragment"],
-    ) -> Optional[List["pyarrow.parquet.FileMetaData"]]:
-        if len(pieces) > PARALLELIZE_META_FETCH_THRESHOLD:
-            return _fetch_metadata_remotely(pieces)
-        else:
-            return _fetch_metadata(pieces)
 
 
 class ParquetDatasource(ParquetBaseDatasource):
